Restructure the code
Try to follow the hexagonal architecture through ops eyes
