+++ internal/application/financialmodelingprep/getResponse.go
