+++ internal/domain/derefstring/derefstring.go
@@ -1,4 +1,4 @@
-package tools
+package derefstring
 
 // DerefString convert pointer string to string
 func DerefString(s *string) string {
